assignments/01a._Data_parsing_server_part1/python/parser/pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["doro-hd <doro-hd@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyyaml = "^6.0.2"
bs4 = "^0.0.2"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
